backend/jd_parser/requirements.txt
openai
python-dotenv
pydantic
beautifulsoup4
